Refactor AnimatedTrain progress calculation to use Duration for improved accuracy; update MainFrame size and split pane location; add speed control slider in MapPanel for dynamic simulation speed adjustment

+++ src/de/drvbund/lernlabit/lb/javaproject/SimulationController.java
@@ -30,7 +30,7 @@ public class SimulationController {
         this.speedFactor = speedFactor;
     }
 
-    public double getSpeedFactor() {
-        return speedFactor;
+    public int getSpeedFactor() {
+        return (int) speedFactor;
     }
 }
